delete update auth

+++ mediablog/views.py
@@ -6,7 +6,7 @@ from django.http import HttpResponse
 from .models import blogcontain
 from .forms import TestFromClass,ModelsDemoForm
 from django.shortcuts import get_object_or_404
-
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 def index(request):
     return render(request,'index.html')
@@ -54,19 +54,34 @@ def modelDjangoFrom(request):
     success = ''
     form = ModelsDemoForm(request.POST,request.FILES or None)
     if form.is_valid():
-        form.save()
+        obj=form.save(commit=False)
+        obj.author=request.user
+        print(request.user)
+        obj.save()
+        
         success = 'Successfully saved data'
 
     context={'form': form, 'success': success}
     return render(request, "pages/model_form.html", context)
-
+@login_required(login_url="/admin/")
 def deleteById(request,id):
-    blogcontain.objects.get(pk= id).delete()
-    success='successfully delted the blog'
+    obj=blogcontain.objects.get(id=id)
+    if obj.author == request.user:
+        blogcontain.objects.get(pk=id).delete()
+        success='successfully delted the blog'
+    else:
+        success=f"you cannot delete this blog author is{obj.author}"
     return render(request,'pages/blogs.html',{'success':success})
 
+@login_required(login_url="/admin/")
 def updateblog(request,id):
+
     obj=blogcontain.objects.get(id=id)
+    if obj.author == request.user:
+        pass
+    else:
+        return HttpResponse("you cannot delete")
+
     return render(request,'pages/update_view.html',{"data":obj})
 
 def updatedata(request,id):
@@ -75,7 +90,7 @@ def updatedata(request,id):
     c_id=id
 
     c_title=request.POST.get('title')
-    c_author=request.POST.get('author')
+    c_author=request.user
     c_description=request.POST.get('description')
     c_no_of_line=request.POST.get('no_of_line')
     print(c_id,c_title)
